chore(cleanup): refactor account-related modules and update import paths
